requirements.txt
autopep8 @ file:///tmp/build/80754af9/autopep8_1596578164842/work
backcall==0.2.0
bcrypt==3.2.0
certifi==2020.11.8
cffi==1.14.3
click==7.1.2
decorator==4.4.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596207638929/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1604101197014/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
itsdangerous==1.1.0
jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2==2.11.2
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1601311786391/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1606148996965/work
MarkupSafe==1.1.1
mysqlclient==2.0.1
parso==0.7.0
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1605560620615/work/dist/ptyprocess-0.6.0-py2.py3-none-any.whl
pycodestyle==2.6.0
pycparser==2.20
Pygments @ file:///tmp/build/80754af9/pygments_1604103097372/work
python-dateutil==2.8.1
pyzmq==20.0.0
six @ file:///tmp/build/80754af9/six_1605205327372/work
SQLAlchemy==1.3.20
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
traitlets @ file:///tmp/build/80754af9/traitlets_1602787416690/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
Werkzeug==1.0.1
